# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crm",
        "description": "crm vue jwt",
        "contact": {
            "email": "info@digex"
        },
        "license": {
            "name": "crm",
            "url": "http://crm.loc/"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "crm  host server"
        }
    ],
    "paths": {
        "/api/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "get user login",
                "description": "desc",
                "operationId": "App\\Http\\Controllers\\AdminController::getAuthenticatedUser",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
...
        "/api/companies/index": {
            "get": {
                "tags": [
                    "companies"
                ],
                "description": "get companies",
                "operationId": "App\\Http\\Controllers\\CompanyController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "pages",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/companies/get": {
            "get": {
                "tags": [
                    "companies"
                ],
                "description": "get company id",
                "operationId": "App\\Http\\Controllers\\CompanyController::get",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "ID of pet to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/settings/currencies/index": {
            "get": {
                "tags": [
                    "settings"
                ],
                "description": "get companies",
                "operationId": "App\\Http\\Controllers\\CurrencyController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/settings/currencies/get": {
            "get": {
                "tags": [
                    "settings"
                ],
                "description": "get companies",
                "operationId": "App\\Http\\Controllers\\CurrencyController::get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id of category",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
...
        "/api/incomes/categories/destroy/{item}": {
            "delete": {
                "tags": [
                    "incomes"
                ],
                "description": "get company by id",
                "operationId": "App\\Http\\Controllers\\InvoiceCategoryController::destroy",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "id of category",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "auth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}